Update docs and remove dataset.extract.
Also move dataset.dataset up a level (so we just have nimare.dataset).
