+++ .gitignore
@@ -1,6 +1,7 @@
 historical/
 docs/_build/
 docs/generated/
+docs/modules/
 docs/auto_examples/
 nimare/resources/data/neurovault-data/
 nimare/tests/data/downloaded/
