+++ nimare/dataset.py
@@ -10,8 +10,8 @@ import numpy as np
 import pandas as pd
 import nibabel as nib
 
-from ..base.base import NiMAREBase
-from ..utils import tal2mni, mni2tal, mm2vox, get_template
+from .base.base import NiMAREBase
+from .utils import tal2mni, mni2tal, mm2vox, get_template
 
 LGR = logging.getLogger(__name__)
