--- nimare/dataset/__init__.py
@@ -1,7 +1,0 @@
-"""
-Top-level namespace for datasets.
-"""
-from .dataset import Dataset
-from .extract import DataSource
-
-__all__ = ['Dataset', 'DataSource']
